tests: Initial raw_transaction integration tests.
